+++ runtime/core/xnix_runtime_core.c
@@ -1995,6 +1995,7 @@ xnix_runtime_compatibility_action_queue(
 #include "xnix_runtime_core_portal_request.inc"
 #include "xnix_runtime_core_compatibility_snapshot_plan.inc"
 #include "xnix_runtime_core_compatibility_test_plan.inc"
+#include "xnix_runtime_core_compatibility_test_result.inc"
 #include "xnix_runtime_core_compatibility_repair_plan.inc"
 #include "xnix_runtime_core_ai_diagnostic_input.inc"
 #include "xnix_runtime_core_ai_diagnostic_recommendation.inc"